ai_face_swap_new/backend/server/requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
grpcio
grpcio-tools
protobuf
numpy
pydantic
python-jose
passlib
bcrypt
python-dotenv
